+++ backend/public/index.php
@@ -52,13 +52,13 @@ define('APP_UPSTREAM', 'beta');
 define('REQUEST_ID', uniqid());
 
 if (APP_DEBUG) {
-	error_reporting(E_ALL);
-	ini_set('display_errors', 1);
-	ini_set('display_startup_errors', 1);
+    error_reporting(E_ALL);
+    ini_set('display_errors', 1);
+    ini_set('display_startup_errors', 1);
 } else {
-	error_reporting(0);
-	ini_set('display_errors', 0);
-	ini_set('display_startup_errors', 0);
+    error_reporting(0);
+    ini_set('display_errors', 0);
+    ini_set('display_startup_errors', 0);
 }
 
 if (APP_DEBUG) {